Add OpenStack support

+++ wolkenbrot/os.py
@@ -35,20 +35,35 @@ class OpenStackBuilder(Builder):
         self.instance = None
         self.ssh_client = None
         self.sec_group_id = None
-        self.sec_group = None
+        self.sec_grp = None
 
     def __exit__(self, exc_type, exc_val, exc_tb):
-        self._shutdown_machine()
-        self._destroy_machine()
-        # TODO: add cleaning of security group
+        self.clean()
+
+    def clean(self):
+        if self.instance:
+            self._shutdown_machine()
+            self._destroy_machine()
+        if self.sec_group_id:
+            print(f"Deleting security groups {self.sec_grp.name}...")
+            self.client.delete_security_group(self.sec_group_id)
+        if self.key:
+            print(f"Deleting keypair {self.key.name}...")
+            self.client.delete_keypair(self.key.id)
+        if self.instance.volumes:
+            print(f"Deleting volume {self.instance.volumes[0].id}")
+            self.client.delete_volume(self.instance.volumes[0].id)
 
     def _shutdown_machine(self):
         print("Shutdown imaging machine...")
-        self.client.compute.stop_server(self.instance.id)
+        instance = self.client.get_server(self.instance.id)
+        if instance and instance.status != 'SHUTOFF':
+            self.client.compute.stop_server(self.instance.id)
 
     def _destroy_machine(self):
-        print("Destroy imaging machine...")
-        self.client.delete_server(self.instance.id, wait=True, timeout=360)
+        if self.client.get_server(self.instance.id):
+            print("Destroy imaging machine...")
+            self.client.delete_server(self.instance.id, wait=True, timeout=360)
 
     def make_new_key(self):
         print("Creating keypair for imaging machine...")
@@ -89,6 +104,8 @@ class OpenStackBuilder(Builder):
             security_groups=self.sec_group_id,
             image=self.image.id,
             key_name=self.key.id,
+            boot_from_volume=True,
+            volume_size=25,
             userdata='manage_etc_hosts: true'
         )
         self.wait_for_status("ACTIVE")
@@ -121,16 +138,14 @@ class OpenStackBuilder(Builder):
         if self.instance.public_v4:
             ip_addr = self.instance.public_v4
         else:
-            ip_addr = self.instance['addresses'][self.config['network']][0]['addr']
+            ip_addr = self.instance['addresses'][self.config['network']["name"]][0]['addr']
 
         print(f"Connecting to {ip_addr} using key {self.key.name}")
-
         for i in range(0, 15):
             try:
-                # TODO: fix hard coded user here
-                client = SSHClient(ip_addr, 22, 'ubuntu', None,
+                self.ssh_client = SSHClient(ip_addr, 22, self.config["user"], None,
                                    self.key.private_key, None)
-                return client
+                return
             except paramiko.ssh_exception.PasswordRequiredException as excep:
                 raise excep
             except (NoValidConnectionsError, TimeoutError,
@@ -146,13 +161,48 @@ class OpenStackBuilder(Builder):
 
     @timeout(1200, "Creating of image took too long ...")
     def create_image(self):
-        pass
+        """
+        This method is responsible of creating the image.
+        """
+        self._shutdown_machine()
+        self._destroy_machine()
+        image_name = self.config['name']
+        print(f'Creating image {image_name}')
+        # delete image for this date if already existing
+        if self.client.get_image(image_name):
+            self.client.delete_image(image_name, wait=True, timeout=360)
 
+        print('Waiting for volume to be available')
+        for i in range(0, 360):
+            status = self.client.get_volume(self.instance.volumes[0].id).status
+            print("Status of volume is {}".format(status))
+            if status == 'available':
+                break
+            time.sleep(10)
+
+        if status != 'available':
+            raise RuntimeError('Volume was has not become available during the allowed'
+                               'time!')
+
+        print('Creating image...')
+        image = self.client.create_image(
+            image_name,
+            wait=True,
+            timeout=3600,
+            disk_format='raw',
+            volume=self.instance.volumes[0].id
+        )
+
+        tags = self.config.get("tags", None)
+        if tags:
+            all_values = {}
+            [all_values.update(tag) for tag in tags]
+            self.client.update_image_properties(image, name=image.name, **all_values)
 
 def list_images(CLIENT):
     for image in CLIENT.list_images():
         print("{id}\t{name:20}\t\t{created}".format(**image))
-    
+
 
 def list_details(CLIENT, image_id):
     image = CLIENT.image.find_image("00fe5e3a-7c97-4071-be12-6ce7d1a5ecf5")
@@ -168,9 +218,10 @@ def list_details(CLIENT, image_id):
             pprint(value)
 
 
-def delete_image(CLIENT, image_id):
-    printr("Deleting ...")
-    CLIENT.image.delete_image(image_id, ignore_missing=True)
+def delete_image(CLIENT, image_name):
+    printr(f"Deleting {image_name} ...")
+    img_id = CLIENT.image.find_image(image_name)
+    CLIENT.image.delete_image(img_id, ignore_missing=True)
 
 
 def validate_image_name(CLIENT, name):
@@ -185,12 +236,10 @@ def bake(CLIENT, image):  # pragma: no coverage
         config_dict = json.load(fd)
 
     check_config(config_dict)
-
     if validate_image_name(CLIENT, config_dict['name']):
         printr("An image named '{}' already exists!!!".format(
             config_dict['name']))
         sys.exit(2)
-
     with OpenStackBuilder(CLIENT, config_dict) as builder:
         builder.launch()
         builder.wait_for_ssh()
